1. Collision 발생할 수 있는 오브젝트간 경우의 수를 비트 연산으로 구현 2. 발생할 수 있는 경우의 수가 많은 것을 방지하기 위한 자료구조 선택

+++ Client/PathManager.cpp
@@ -16,10 +16,10 @@ void PathManager::Init()
 	// 현재 경로 띄워준다.
 	GetCurrentDirectory(255, _contentDirPath);
 	
-	int len = wcslen(_contentDirPath);
+	size_t len = wcslen(_contentDirPath);
 
 #ifdef _DEBUG												
-	for (int i = len - 1; i >= 0; --i)	
+	for (size_t i = len - 1; i >= 0; --i)	
 	{
 		if (_contentDirPath[i] == '\\')					
 		{												
@@ -29,7 +29,7 @@ void PathManager::Init()
 	}													
 #else														
 	int cnt = 0;										
-	for (int i = len - 1; i >= 0; --i)					
+	for (size_t i = len - 1; i >= 0; --i)
 	{													
 		if (cnt == 0 && _contentDirPath[i] == '\\')		
 		{
